refactor: restructure CPF CLI tool and remove unused files
- Moved main application logic to cmd/cpf/main.go for better organization.
- Created a new package 'pkg/cpf' to encapsulate CPF-related functionalities.
- Removed obsolete main.go and cpf_test.go files.
- Updated module path in go.mod to reflect new repository structure.
- Enhanced .gitignore to exclude unnecessary files.
- Adjusted goreleaser configuration for the new structure.
