+++ .gitignore
@@ -1,5 +1,6 @@
 # Binaries
 cpf
+!cpf/
 cpf-cli-go
 dist/
