+++ .goreleaser.yaml
@@ -4,7 +4,8 @@ before:
     - go mod tidy
 
 builds:
-  - env:
+  - main: ./cmd/cpf
+    env:
       - CGO_ENABLED=0
     goos:
       - darwin
