+++ go.mod
@@ -1,3 +1,3 @@
-module cpf-cli-go
+module github.com/diegopeixoto/cpf-cli-go
 
 go 1.23
